fix error

+++ src/scenes/PhotoScene.js
@@ -348,6 +348,9 @@ class PhotoScene extends Component {
         if (photoUUID) {
             try {
                 const oldPhoto = this.props.photos.find(photo => photo.uuid === photoUUID);
+                if (oldPhoto === undefined) {
+                    throw new Error("oldPhoto.uri not exists");
+                }
                 const result = await this.props.deleteImage(oldPhoto.uri, oldPhoto.id, true);
                 if (result) {
                     await unlink(getPhotoPath(oldPhoto.uri));
